feat: move `Accumulator` to `accumulator.rs`
